animal-avatar-orchestrator/requirements.txt
ollama>=0.1.0
torch>=2.0.0
diffusers>=0.20.0
transformers>=4.30.0
accelerate>=0.21.0
Pillow>=10.0.0
python-dotenv>=1.0.0
